Don't use the root property to determine suite root.
The root property is problematic because Mocha sets it true if the
description of a suite is the empty string. However, Mocha runs
perfectly well with suites that have empty names. See:

https://github.com/mochajs/mocha/issues/2755

Instead of relying on the root property, rely on whether the parent
property is set. If not, then we are at the root.